requirements.txt
dj_database_url==0.5.0
Django==3.1.8
django_filter==2.4.0
django_heroku==0.3.1
django-extensions
django-widget-tweaks
django-tempus-dominus
langdetect
Faker